requirements.txt
streamlit
groq
python-docx
PyPDF2
pdfplumber
reportlab